Ajout d'un mémo en DB après appui sur le bouton "save memo"

+++ app/src/main/java/ephec/noticeme/Alarm.java
@@ -12,8 +12,8 @@
         private String modificationDate;
         private String title;
         private String description;
-        private float latitude;
-        private float longitude;
+        private double latitude;
+        private double longitude;
         private String alarmDate;
 
         public Alarm() {
@@ -46,19 +46,19 @@
             this.alarmDate = alarmDate;
         }
 
-        public float getLongitude() {
+        public double getLongitude() {
             return longitude;
         }
 
-        public void setLongitude(float longitude) {
+        public void setLongitude(double longitude) {
             this.longitude = longitude;
         }
 
-        public float getLatitude() {
+        public double getLatitude() {
             return latitude;
         }
 
-        public void setLatitude(float latitude) {
+        public void setLatitude(double latitude) {
             this.latitude = latitude;
         }
